tests/monorepo-05/apps/processor/requirements.txt
pyyaml==5.3.1
redis==3.5.3
celery==4.4.7
